igvf-catalog-llm/requirements.txt
Flask==2.2.3
python-arango==7.5.7
langchain-community==0.3.31
langchain-openai==0.3.35
langchain-arangodb==1.2.0
Werkzeug==2.2.2
gunicorn==23.0.0
Flask-Limiter==3.11.0
pre-commit==4.2.0
